+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 - Added download QR code page to not generated file to be signed.
   [chris-adam, sgeulette]
+- Added signer substitutes table.
+  [chris-adam]
+
 
 3.1.1 (2026-03-27)
 ------------------